backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
supabase==1.0.4
openai==1.3.7
groq==0.11.0
pdfplumber==0.10.3
pytesseract==0.3.10
pillow==10.1.0
scikit-learn==1.3.2
lightgbm==4.6.0
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
slowapi==0.1.8
# httpx==0.25.2